requirements.txt
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==4.4.2
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.53.1
h5py==3.11.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
neo==0.13.3
nest-asyncio==1.6.0
npTDMS==1.9.0
numpy==1.26.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
Pillow==8.3.2
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
PyQt5
PyQt5-Qt5
PyQt5_sip
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
qtconsole==5.6.0
QtPy==2.4.1
quantities==0.15.0
requests==2.32.3
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
tifffile==2024.9.20
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
python-vlc
